django_project_6/itcompany/requirements.txt
Django==5.2
Pillow==12.1.0
djangorestframework==3.15.0
django-filter==23.5
django-cors-headers==4.3.1
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1